+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 
 ## [Unreleased]
 ### Added
+- markdown rendering
 - link to changelog in docs panel
 - progress bar to show simulation state
 - WMS layers flurstuecke and ertragspotenzial